+++ libs/mifos-loans/src/commonMain/kotlin/org/mifos/lib/loan/core/model/Loan.kt
@@ -7,13 +7,16 @@
  *
  * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
  */
-package org.mifospay.mifos.loans.core.model
+package org.mifos.lib.loan.core.model
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LoanPayment(
-    val loanId: Long,
-    val amount: Double,
-    val paymentDate: String,
+data class Loan(
+    val id: Long,
+    val productId: Long,
+    val principal: Double,
+    val interestRate: Double,
+    val status: String,
+    val schedule: LoanSchedule? = null,
 )
