+++ libs/mifos-loans/src/commonMain/kotlin/org/mifos/lib/loan/core/model/LoanProduct.kt
@@ -7,7 +7,7 @@
  *
  * See https://github.com/openMF/mobile-wallet/blob/master/LICENSE.md
  */
-package org.mifospay.mifos.loans.core.model
+package org.mifos.lib.loan.core.model
 
 import kotlinx.serialization.Serializable
